backend/pyproject.toml
[project]
name = "npa-2024-todoapp"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "hanabi", email = "r.a.n.a.b.1.5103@gmail.com" }
]
dependencies = [
    "django-ninja>=1.3.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/npa_2024_todoapp"]
